Move test cases to fixtures
