+++ modules/bazi-calculator-manual/src/test/kotlin/net/twisterrob/astro/bazi/ManualCalculatorTest.kt
@@ -2,8 +2,6 @@ package net.twisterrob.astro.bazi
 
 import net.twisterrob.astro.bazi.model.BaZi
 import org.junit.jupiter.api.Assumptions
-import org.junit.jupiter.api.DynamicNode
-import org.junit.jupiter.api.TestFactory
 import java.time.LocalDateTime
 
 class ManualCalculatorTest : BaZiCalculatorTest() {
@@ -23,28 +21,4 @@ class ManualCalculatorTest : BaZiCalculatorTest() {
 			}
 		}
 	}
-
-	@TestFactory fun `random tests`(): Iterable<DynamicNode> =
-		BaZiTestCase.ALL_CASES.map { subject.verify(it.dateTime, it.bazi) }
-
-	@TestFactory fun `random hourless tests`(): Iterable<DynamicNode> =
-		BaZiHourlessTestCase.ALL_CASES.map { subject.verify(it.date, it.bazi) }
-
-	@TestFactory fun `sexagenary years`(): Iterable<DynamicNode> =
-		SexagenaryYearTestCase.ALL_KNOWN_CYCLES.map { subject.verifyCycle(it) }
-
-	@TestFactory fun `solar term branches of non-leap year`(): Iterable<DynamicNode> =
-		listOf(2018, 2022).map { year -> subject.verifySolarTermBranchesIn(year) }
-
-	@TestFactory fun `solar term stems of non-leap years`(): Iterable<DynamicNode> =
-		(1900..2100).map { year -> subject.verifySolarTermStemsIn(year) }
-
-	@TestFactory fun `sexagenary days`(): Iterable<DynamicNode> =
-		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.map { subject.verifyCycle(it) }
-
-	@TestFactory fun `special days`(): Iterable<DynamicNode> =
-		SexagenaryDayTestCase.WIKIPEDIA_EXAMPLES.map { subject.verifyDay(it) }
-
-	@TestFactory fun `sexagenary hours`(): Iterable<DynamicNode> =
-		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.map { subject.verifyCycle(it) }
 }
