+++ modules/bazi-calculator/src/main/kotlin/net/twisterrob/astro/bazi/lookup/HeavenlyStem_lookupHour.kt
