+++ modules/bazi-calculator/src/testFixtures/kotlin/net/twisterrob/astro/bazi/test/data/BaZiTestCase.kt
@@ -1,4 +1,4 @@
-package net.twisterrob.astro.bazi
+package net.twisterrob.astro.bazi.test.data
 
 import net.twisterrob.astro.bazi.model.BaZi
 import net.twisterrob.astro.bazi.model.EarthlyBranch.Chen
