+++ modules/bazi-calculator/src/testFixtures/kotlin/net/twisterrob/astro/bazi/test/data/SexagenaryDayTestCaseTest.kt
@@ -1,4 +1,4 @@
-package net.twisterrob.astro.bazi
+package net.twisterrob.astro.bazi.test.data
 
 import io.kotest.assertions.withClue
 import io.kotest.matchers.collections.shouldHaveSize
@@ -8,11 +8,11 @@ import net.twisterrob.astro.bazi.lookup.sexagenaryCycle
 import net.twisterrob.astro.bazi.model.BaZi
 import org.junit.jupiter.api.Test
 
-class SexagenaryYearTestCaseTest {
+class SexagenaryDayTestCaseTest {
 
 	@Test
 	fun `the test data has full cycles`() {
-		SexagenaryYearTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle shouldHaveSize 60
 			cycle.forEachIndexed { index, testCase ->
 				testCase.cyclicOrdinal shouldBe index + 1
@@ -22,10 +22,10 @@ class SexagenaryYearTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with cycles`() {
-		SexagenaryYearTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle.forEach { testCase ->
 				val expectedPillar = BaZi.Pillar.sexagenaryCycle(testCase.cyclicOrdinal - 1)
-				val actualPillar = BaZi.Pillar(testCase.stem, testCase.branch)
+				val actualPillar = BaZi.Pillar(testCase.dayStem, testCase.dayBranch)
 
 				withClue(testCase) {
 					actualPillar shouldBe expectedPillar
@@ -36,9 +36,10 @@ class SexagenaryYearTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with dates`() {
-		SexagenaryYearTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
-			cycle.forEach { testCase ->
-				testCase.startDate shouldBeBefore testCase.endDate
+		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+			cycle.take(59).zip(cycle.drop(1)).forEach { (testCase, nextTestCase) ->
+				testCase.date shouldBeBefore nextTestCase.date
+				nextTestCase.date shouldBe testCase.date.plusDays(1)
 			}
 		}
 	}
