+++ modules/bazi-calculator/src/testFixtures/kotlin/net/twisterrob/astro/bazi/test/data/SexagenaryHourTestCaseTest.kt
@@ -1,18 +1,19 @@
-package net.twisterrob.astro.bazi
+package net.twisterrob.astro.bazi.test.data
 
 import io.kotest.assertions.withClue
 import io.kotest.matchers.collections.shouldHaveSize
+import io.kotest.matchers.date.shouldBeAfter
 import io.kotest.matchers.date.shouldBeBefore
 import io.kotest.matchers.shouldBe
 import net.twisterrob.astro.bazi.lookup.sexagenaryCycle
 import net.twisterrob.astro.bazi.model.BaZi
 import org.junit.jupiter.api.Test
 
-class SexagenaryDayTestCaseTest {
+class SexagenaryHourTestCaseTest {
 
 	@Test
 	fun `the test data has full cycles`() {
-		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle shouldHaveSize 60
 			cycle.forEachIndexed { index, testCase ->
 				testCase.cyclicOrdinal shouldBe index + 1
@@ -22,10 +23,10 @@ class SexagenaryDayTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with cycles`() {
-		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
 			cycle.forEach { testCase ->
 				val expectedPillar = BaZi.Pillar.sexagenaryCycle(testCase.cyclicOrdinal - 1)
-				val actualPillar = BaZi.Pillar(testCase.dayStem, testCase.dayBranch)
+				val actualPillar = BaZi.Pillar(testCase.hourStem, testCase.hourBranch)
 
 				withClue(testCase) {
 					actualPillar shouldBe expectedPillar
@@ -36,10 +37,11 @@ class SexagenaryDayTestCaseTest {
 
 	@Test
 	fun `the test data is consistent with dates`() {
-		SexagenaryDayTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
-			cycle.take(59).zip(cycle.drop(1)).forEach { (testCase, nextTestCase) ->
-				testCase.date shouldBeBefore nextTestCase.date
-				nextTestCase.date shouldBe testCase.date.plusDays(1)
+		SexagenaryHourTestCase.ALL_KNOWN_CYCLES.forEach { cycle ->
+			cycle.forEach { testCase ->
+				testCase.startTime shouldBeBefore testCase.endTime
+				testCase.midTime shouldBeAfter testCase.startTime
+				testCase.midTime shouldBeBefore testCase.endTime
 			}
 		}
 	}
